prod

+++ src/abstraction/eq1Producteur1/Producteur1arbes.java
@@ -10,16 +10,12 @@ public class Producteur1arbes extends plantation {
     private plantation moyenne_qualite;
     private plantation haute_qualite;
     
-
-    
-
-
     public Producteur1arbes(plantation basse_qualite, plantation moyenne_qualite, plantation haute_qualite){
         super();
         HashMap<plantation, Double> map = new HashMap<>();
-        map.put(this.basse_qualite(), 25);
-        map.put(this.moyenne_qualite(), 30);
-        map.put(this.haute_qualite(), 20);
+        map.put(this.basse_qualite, (double) 25);
+        map.put(this.moyenne_qualite, (double) 30);
+        map.put(this.haute_qualite, (double) 20);
     }
